added more functionality and simplified unzip

+++ src/helpers.py
@@ -4,13 +4,23 @@ from zipfile import ZipFile
 import tensorflow as tf
 
 
-def extractzip(zippedfilename, outputfoldername):
-    if os.path.isdir(outputfoldername):
-        shutil.rmtree(outputfoldername)
-    with ZipFile(zippedfilename, 'r') as zipObj:
-       # Extract all the contents of zip file in different directory
-       zipObj.extractall()
-    os.rename(zippedfilename.split('.')[0], outputfoldername)
+def unzip(fname,oname=None):
+    with ZipFile(fname, 'r') as zipObj:
+        if oname is not None:
+            zipObj.extractall(oname)
+        else:
+            zipObj.extractall()
+        print('extracted to: %s' % fname)
+
+
+def extractzip(zippedfilename, outputfoldername=None):
+    if outputfoldername is not None:
+        if os.path.isdir(outputfoldername):
+            shutil.rmtree(outputfoldername)
+            print('remove already existing folder: %s' % outputfoldername)
+        unzip(zippedfilename, outputfoldername)
+    else:
+        unzip(zippedfilename)
 
 
 def check_gpus():
